Final tweak to anchor stuff
After some heartache, here is the story about anchors.

The ultimate HTML for anchored links is:

```html
<hX class="j2-anchor" id="heading-id">
 <span data-anchor-id="heading-id">Heading Name</span>
</hX>
```

With anchor setup:
```js
anchors.options.visible = touch
anchors.add('.j2-anchor span')
```

Here are the constraints:
1. Anchor.js places its anchor as a child of some selector.
   It generates the `href` in the link from that element's `id`
   (or, indirect via `data-anchor-id`.)

2. We want to use hover, meaning anchor sets a style for
   `*:hover > anchor-js-link` to display itself.

3. We have a fixed heading, and use the `::before` trick to
   insert an invisible fixed-height element before the anchor
   target.

Now the problem is that #2 and #3 collide.  If the ::before
pseudo-element is hovered over, that triggers the #2 selector,
so the `::before` is treated for '>' purposes as part of its
creating element rather than a sibling.  Unexpected!  Looks bad.

So, we solve this by splitting the link-target element (needs
the `::before`) from the anchor.js element.  And luckily the
`data-anchor-id` thing is there to complete the circuit.